Add files via upload
This project consists of code in SQL used to find the best smartphone by brand based on website data from 2019 and displaying the results in Tableau. To run, please download the repository, run the 'Business Data.sql' file, and import the 'Business Data-'minimum price for smartphones of each brand for oct2019'' and the 'Business Data-'minimum price for smartphones of each brand for nov2019'' into the SQL program you use to view the project.